Refactored content from App component to Pokedex Search Component
