+++ pokedex-react-app/src/components/App/App.css
@@ -11,12 +11,3 @@
     }
 }
 
-.arrows-ui{
-    z-index: 3;
-    position: sticky;
-    top: 50vh;
-    left: 5%;
-    display: flex;
-    width: 90%;
-    justify-content: space-between;
-}
